+++ docs/release_notes.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## Version 1.1.2 (December 2, 2022)
+
+* Added support for python version 3.6 or later
+* Added georeferencing for geotiff files when using the create_mosaic function
+
 ## Version 1.1.1 (November 21, 2022)
 
 * Added `filename` parameter to create_mosaic and get_full_res_image functions